Rewrote from ground up, now works
First fixed the get col method, my compare statement was wrong way round so it always skipped to LHS. Now can write to both sides. Deleted extraneous code in GLCD_read. Fixed the mod and div operations. Added a mod_to_pattern which takes a modulus value and converts that to a binary number with the kth bit on. (needed because bsf doesn't work with a file register). Adjusted the clear screen method to work properly.