External LCD display code
Added subroutines to display the brush size and mode of drawing.
Added subroutines W_Brushsize and W_Colour to move the brush size and colour variables to W, to determine display message.
Default display subroutine is update_display (should be called whenever brush size is changed). sending_display and receiving_display are for whenever an image is sent or received from UART.
Need to implement these with rest of the program, and test with keypad inputs.